Use csv format for konj.cache (breaking change)
Note that this change is NOT backwards compatible. New versions
of konj will not be able to read old konj.cache files. Automated
conversion of old konj.cache files to the new format would be
possible but has not been implemented. If you have old konj.cache
files anywhere, make sure to convert them to the new format
yourself; or simply delete them and start over without a cache.